Fix int64->int32 warning
git-svn-id: http://skia.googlecode.com/svn/trunk@5330 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ include/gpu/GrTypes.h
@@ -115,7 +115,7 @@ static inline size_t GrSizeAlignUpPad(size_t x, size_t alignment) {
 static inline uint32_t GrUIAlignDown(uint32_t x, uint32_t alignment) {
     return (x / alignment) * alignment;
 }
-static inline uint32_t GrSizeAlignDown(size_t x, uint32_t alignment) {
+static inline size_t GrSizeAlignDown(size_t x, uint32_t alignment) {
     return (x / alignment) * alignment;
 }
